style: post page 폴더구조 변경
