+++ src/app/post/[nid]/page.tsx
@@ -1,1 +1,1 @@
-export { default } from "@components/pages/Post";
+export { default } from "@components/pages/Post/[nid]";
