+++ src/components/pages/Post/index.tsx
@@ -1,27 +1,36 @@
 import React from "react";
 
-import PostHeader from "@components/ui/Post/PostHeader";
-import PostFooter from "@components/ui/Post/PostFooter";
-import PostContent from "@components/ui/Post/PostContent";
+import { USER_DETAIL, USER_LINKS } from "@utils/constant/user";
+import HomeContent from "@components/ui/HomeContent";
+import Avatar from "@components/ui/core/Avatar";
 import Hydrate from "@components/pages/Post/Hydrate";
+import { IGetPostsInput } from "@/types";
 
 interface IProps {
-  params: {
-    nid: number;
-  };
+  searchParams: IGetPostsInput;
 }
 
-const Post = ({ params }: IProps) => {
-  const nid = Number(params.nid);
-
+const Post = async ({ searchParams }: IProps) => {
   return (
-    <section className="frame flex w-[theme(screens.xl.max)] flex-col items-center xl:w-full">
-      <Hydrate nid={nid}>
-        <PostHeader nid={nid} />
-        <PostContent nid={nid} />
-        <PostFooter nid={nid} />
+    <main className="frame flex flex-col items-center justify-center">
+      <section className="3-16 flex w-[700px] items-center justify-center gap-7 px-0 py-16 text-primary md:w-full md:flex-col md:gap-4 md:px-0 md:pb-12 md:pt-8">
+        <Avatar width={140} height={140} />
+        <div className="flex flex-1 flex-col gap-4 transition-[color]">
+          <h3 className="text-xl font-medium">👨🏻‍💻 {USER_DETAIL.username}</h3>
+          <p className="font-light">{USER_DETAIL.description}</p>
+          <div className="flex w-full items-center gap-5">
+            {USER_LINKS.map(({ name, link, Icon }) => (
+              <a key={name} href={link} target="_blank" rel="noopener noreferrer" aria-label={name}>
+                <Icon className="h-6 w-6 fill-[theme(textColor.secondary)] transition-[fill] hover:fill-[theme(textColor.effect1)]" />
+              </a>
+            ))}
+          </div>
+        </div>
+      </section>
+      <Hydrate params={searchParams}>
+        <HomeContent params={searchParams} />
       </Hydrate>
-    </section>
+    </main>
   );
 };
